* grafana/master: (54 commits)
  Plugins: update beta notice style (#16928)
  Chore: update version number for 6.3 (#16927)
  Plugins: Support templated urls in routes (#16599)
  changelog: add 5.4.4 release
  docs: add download link to what's new in v6.2
  update changelog
  Update changelog for 6.2.0-beta1
  AzureMonitor: adds support for multiple subscriptions per datasource (#16922)
  docs: what's new in v6.2 (#16909)
  Chore: ban importing from @grafana/ui in grafana ui files (#16920)
  BarGauge: New multi series enabled gauge like panel with horizontal and vertical layouts and 3 display modes (#16918)
  alerting: no notification when going from nodata -> pending (#16905)
  rpm: start grafana after mysqld process (#16917)
  Build: fix failing grafana/ui build (#16913)
  BarGauge: Updated test dashboards and threshold fix (#16911)
  PanelModel: Clear queryRunner on destroy (#16906)
  Panels: No title will no longer make panel header take up space (#16884)
  Elasticsearch: Fix pre-v7.0 and alerting error (#16904)
  Gauge: Better handling of gauge repeat title (#16886)
  Refactor: move datemath to grafana/ui (#16890)
  ...

+++ pkg/tsdb/azuremonitor/azuremonitor-datasource_test.go
@@ -107,19 +107,19 @@ func TestAzureMonitorDatasource(t *testing.T) {
 				So(len(res.Series[0].Points), ShouldEqual, 5)
 
 				So(res.Series[0].Points[0][0].Float64, ShouldEqual, 2.0875)
-				So(res.Series[0].Points[0][1].Float64, ShouldEqual, 1549620780000)
+				So(res.Series[0].Points[0][1].Float64, ShouldEqual, int64(1549620780000))
 
 				So(res.Series[0].Points[1][0].Float64, ShouldEqual, 2.1525)
-				So(res.Series[0].Points[1][1].Float64, ShouldEqual, 1549620840000)
+				So(res.Series[0].Points[1][1].Float64, ShouldEqual, int64(1549620840000))
 
 				So(res.Series[0].Points[2][0].Float64, ShouldEqual, 2.155)
-				So(res.Series[0].Points[2][1].Float64, ShouldEqual, 1549620900000)
+				So(res.Series[0].Points[2][1].Float64, ShouldEqual, int64(1549620900000))
 
 				So(res.Series[0].Points[3][0].Float64, ShouldEqual, 3.6925)
-				So(res.Series[0].Points[3][1].Float64, ShouldEqual, 1549620960000)
+				So(res.Series[0].Points[3][1].Float64, ShouldEqual, int64(1549620960000))
 
 				So(res.Series[0].Points[4][0].Float64, ShouldEqual, 2.44)
-				So(res.Series[0].Points[4][1].Float64, ShouldEqual, 1549621020000)
+				So(res.Series[0].Points[4][1].Float64, ShouldEqual, int64(1549621020000))
 			})
 
 			Convey("when data from query aggregated as total to one time series", func() {
@@ -139,7 +139,7 @@ func TestAzureMonitorDatasource(t *testing.T) {
 				So(err, ShouldBeNil)
 
 				So(res.Series[0].Points[0][0].Float64, ShouldEqual, 8.26)
-				So(res.Series[0].Points[0][1].Float64, ShouldEqual, 1549718940000)
+				So(res.Series[0].Points[0][1].Float64, ShouldEqual, int64(1549718940000))
 			})
 
 			Convey("when data from query aggregated as maximum to one time series", func() {
@@ -159,7 +159,7 @@ func TestAzureMonitorDatasource(t *testing.T) {
 				So(err, ShouldBeNil)
 
 				So(res.Series[0].Points[0][0].Float64, ShouldEqual, 3.07)
-				So(res.Series[0].Points[0][1].Float64, ShouldEqual, 1549722360000)
+				So(res.Series[0].Points[0][1].Float64, ShouldEqual, int64(1549722360000))
 			})
 
 			Convey("when data from query aggregated as minimum to one time series", func() {
@@ -179,7 +179,7 @@ func TestAzureMonitorDatasource(t *testing.T) {
 				So(err, ShouldBeNil)
 
 				So(res.Series[0].Points[0][0].Float64, ShouldEqual, 1.51)
-				So(res.Series[0].Points[0][1].Float64, ShouldEqual, 1549723380000)
+				So(res.Series[0].Points[0][1].Float64, ShouldEqual, int64(1549723380000))
 			})
 
 			Convey("when data from query aggregated as Count to one time series", func() {
@@ -199,7 +199,7 @@ func TestAzureMonitorDatasource(t *testing.T) {
 				So(err, ShouldBeNil)
 
 				So(res.Series[0].Points[0][0].Float64, ShouldEqual, 4)
-				So(res.Series[0].Points[0][1].Float64, ShouldEqual, 1549723440000)
+				So(res.Series[0].Points[0][1].Float64, ShouldEqual, int64(1549723440000))
 			})
 
 			Convey("when data from query aggregated as total and has dimension filter", func() {
